# swagger/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TodoList API",
    "description": "카테고리 분류와 종료일 기반 상태 구분을 지원하는 개인 할일 관리 애플리케이션의 REST API입니다.\n\n## 인증\n모든 보호된 엔드포인트는 `Authorization: Bearer <token>` 헤더를 요구합니다. 토큰은 `/api/auth/login` 또는 `/api/auth/register`에서 발급받습니다.\n\n## 할일 상태 판별 로직\n상태(status)는 DB에 별도 저장되지 않으며 조회 시점에 동적으로 계산됩니다.\n- `is_completed = true` → `completed`\n- `is_completed = false` AND (`due_date IS NULL` OR `due_date >= NOW()`) → `pending`\n- `is_completed = false` AND `due_date < NOW()` → `overdue`\n\n## 공통 응답 형식\n- 성공: `{ \"success\": true, \"data\": ... }`\n- 오류: `{ \"success\": false, \"message\": \"...\" }`",
    "version": "1.0.0",
    "contact": {
      "name": "soominlee",
      "email": "26sh.digital@gmail.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "로컬 개발 서버"
    }
  ],
...
  "paths": {
...
    "/api/categories": {
      "get": {
        "tags": ["Categories"],
        "summary": "카테고리 목록 조회",
        "description": "현재 인증된 사용자의 카테고리 목록을 생성일 오름차순으로 반환합니다. 다른 사용자의 카테고리는 조회되지 않습니다.",
        "operationId": "getCategories",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "카테고리 목록 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/SuccessResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "categories": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/Category"
                              }
                            }
                          }
                        }
                      }
                    }
                  ]
                },
                "example": {
                  "success": true,
                  "data": {
                    "categories": [
                      {
                        "id": 1,
                        "userId": 1,
                        "name": "업무",
                        "createdAt": "2026-04-28T00:00:00.000Z"
                      },
                      {
                        "id": 2,
                        "userId": 1,
                        "name": "개인",
                        "createdAt": "2026-04-28T01:00:00.000Z"
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "tags": ["Categories"],
        "summary": "카테고리 생성",
        "description": "새 카테고리를 생성합니다. 카테고리 이름은 필수이며 최대 20자입니다. 동일 사용자 내에서 이름이 중복될 수 없습니다.",
        "operationId": "createCategory",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryCreateRequest"
              },
              "example": {
                "name": "업무"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "카테고리 생성 성공",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/SuccessResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "category": {
                              "$ref": "#/components/schemas/Category"
                            }
                          }
                        }
                      }
                    }
                  ]
                },
                "example": {
                  "success": true,
                  "data": {
                    "category": {
                      "id": 1,
                      "userId": 1,
                      "name": "업무",
                      "createdAt": "2026-04-28T00:00:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "입력값 유효성 검사 실패 (이름 누락 또는 20자 초과)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": {
                  "success": false,
                  "message": "카테고리 이름은 필수이며 최대 20자입니다."
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "409": {
            "description": "동일 사용자 내 카테고리 이름 중복",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": {
                  "success": false,
                  "message": "이미 존재하는 카테고리 이름입니다."
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/categories/{id}": {
      "put": {
        "tags": ["Categories"],
        "summary": "카테고리 수정",
        "description": "카테고리 이름을 수정합니다. 본인 소유의 카테고리만 수정 가능합니다. 수정 후 이름이 동일 사용자 내 다른 카테고리와 중복될 수 없습니다.",
        "operationId": "updateCategory",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "$ref": "#/components/parameters/CategoryId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryCreateRequest"
              },
              "example": {
                "name": "개인 프로젝트"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "카테고리 수정 성공",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/SuccessResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "category": {
                              "$ref": "#/components/schemas/Category"
                            }
                          }
                        }
                      }
                    }
                  ]
                },
                "example": {
                  "success": true,
                  "data": {
                    "category": {
                      "id": 1,
                      "userId": 1,
                      "name": "개인 프로젝트",
                      "createdAt": "2026-04-28T00:00:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "입력값 유효성 검사 실패 (이름 누락 또는 20자 초과)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": {
                  "success": false,
                  "message": "카테고리 이름은 필수이며 최대 20자입니다."
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/CategoryNotFound"
          },
          "409": {
            "description": "동일 사용자 내 카테고리 이름 중복",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": {
                  "success": false,
                  "message": "이미 존재하는 카테고리 이름입니다."
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "tags": ["Categories"],
        "summary": "카테고리 삭제",
        "description": "카테고리를 영구 삭제합니다. 본인 소유의 카테고리만 삭제 가능합니다. 삭제 시 해당 카테고리에 속한 할일의 `categoryId`는 `null`(미분류)로 변경되며, 할일 자체는 삭제되지 않습니다.",
        "operationId": "deleteCategory",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "$ref": "#/components/parameters/CategoryId"
          }
        ],
        "responses": {
          "204": {
            "description": "카테고리 삭제 성공 (응답 본문 없음)"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/CategoryNotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/todos": {
      "get": {
        "tags": ["Todos"],
        "summary": "할일 목록 조회",
        "description": "현재 인증된 사용자의 할일 목록을 조회합니다. `status`와 `categoryId` 쿼리 파라미터로 필터링할 수 있습니다. 상태(status)는 DB에 저장되지 않으며 `is_completed`와 `due_date`를 기반으로 동적 계산됩니다.",
        "operationId": "getTodos",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "할일 상태 필터. 미입력 시 전체 조회.",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/TodoStatus"
            }
          },
          {
            "name": "categoryId",
            "in": "query",
            "description": "카테고리 ID 필터. 미입력 시 전체 조회. `0` 또는 `null`을 전달하면 미분류 할일만 조회할 수 있습니다.",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "할일 목록 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/SuccessResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "todos": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/Todo"
                              }
                            }
                          }
                        }
                      }
                    }
                  ]
                },
                "example": {
                  "success": true,
                  "data": {
                    "todos": [
                      {
                        "id": 1,
                        "userId": 1,
                        "categoryId": 1,
                        "title": "API 설계 완료하기",
                        "description": "OpenAPI 3.0 스펙 작성",
                        "dueDate": "2026-04-30T18:00:00.000Z",
                        "isCompleted": false,
                        "status": "pending",
                        "createdAt": "2026-04-28T00:00:00.000Z",
                        "updatedAt": "2026-04-28T00:00:00.000Z"
                      }
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "쿼리 파라미터 유효성 검사 실패 (잘못된 status 값 등)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": {
                  "success": false,
                  "message": "status는 pending, completed, overdue 중 하나여야 합니다."
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "tags": ["Todos"],
        "summary": "할일 생성",
        "description": "새 할일을 생성합니다. 제목은 필수(최대 100자)이며, 설명(최대 500자), 카테고리, 종료일은 선택 사항입니다. 생성 시 완료 여부는 기본값 `false`(미완료)로 설정됩니다.",
        "operationId": "createTodo",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoCreateRequest"
              },
              "example": {
                "title": "API 설계 완료하기",
                "description": "OpenAPI 3.0 스펙 작성",
                "categoryId": 1,
                "dueDate": "2026-04-30T18:00:00.000Z"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "할일 생성 성공",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/SuccessResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "todo": {
                              "$ref": "#/components/schemas/Todo"
                            }
                          }
                        }
                      }
                    }
                  ]
                },
                "example": {
                  "success": true,
                  "data": {
                    "todo": {
                      "id": 1,
                      "userId": 1,
                      "categoryId": 1,
                      "title": "API 설계 완료하기",
                      "description": "OpenAPI 3.0 스펙 작성",
                      "dueDate": "2026-04-30T18:00:00.000Z",
                      "isCompleted": false,
                      "status": "pending",
                      "createdAt": "2026-04-28T00:00:00.000Z",
                      "updatedAt": "2026-04-28T00:00:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "입력값 유효성 검사 실패 (제목 누락, 100자 초과, 설명 500자 초과 등)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": {
                  "success": false,
                  "message": "제목은 필수이며 최대 100자입니다."
                }
              }
            }
          },
          "400": {
            "description": "지정한 categoryId에 해당하는 카테고리가 존재하지 않거나 본인 소유가 아님",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": {
                  "success": false,
                  "message": "존재하지 않거나 접근할 수 없는 카테고리입니다."
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/todos/{id}": {
      "put": {
        "tags": ["Todos"],
        "summary": "할일 수정",
        "description": "할일의 제목, 설명, 카테고리, 종료일을 수정합니다. 본인 소유의 할일만 수정 가능합니다. 요청 본문에 포함된 필드만 업데이트됩니다.",
        "operationId": "updateTodo",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "$ref": "#/components/parameters/TodoId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoUpdateRequest"
              },
              "example": {
                "title": "API 설계 및 문서화 완료하기",
                "description": "OpenAPI 3.0 스펙 작성 및 Swagger UI 연동",
                "categoryId": 1,
                "dueDate": "2026-04-30T20:00:00.000Z",
                "isCompleted": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "할일 수정 성공",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/SuccessResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "todo": {
                              "$ref": "#/components/schemas/Todo"
                            }
                          }
                        }
                      }
                    }
                  ]
                },
                "example": {
                  "success": true,
                  "data": {
                    "todo": {
                      "id": 1,
                      "userId": 1,
                      "categoryId": 1,
                      "title": "API 설계 및 문서화 완료하기",
                      "description": "OpenAPI 3.0 스펙 작성 및 Swagger UI 연동",
                      "dueDate": "2026-04-30T20:00:00.000Z",
                      "isCompleted": false,
                      "status": "pending",
                      "createdAt": "2026-04-28T00:00:00.000Z",
                      "updatedAt": "2026-04-28T12:00:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "입력값 유효성 검사 실패 (제목 100자 초과, 설명 500자 초과 등)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": {
                  "success": false,
                  "message": "제목은 최대 100자입니다."
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/TodoNotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "tags": ["Todos"],
        "summary": "할일 삭제",
        "description": "할일을 영구 삭제합니다. 본인 소유의 할일만 삭제 가능합니다. 삭제된 데이터는 복구할 수 없습니다.",
        "operationId": "deleteTodo",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "$ref": "#/components/parameters/TodoId"
          }
        ],
        "responses": {
          "204": {
            "description": "할일 삭제 성공 (응답 본문 없음)"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/TodoNotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT 액세스 토큰. 알고리즘: HS-512. `/api/auth/login` 또는 `/api/auth/register`에서 발급. `Authorization: Bearer <token>` 형식으로 전달."
      }
    },
    "schemas": {
...
      "Category": {
        "type": "object",
        "description": "카테고리 정보",
        "properties": {
          "id": {
            "type": "integer",
            "description": "카테고리 고유 식별자 (자동 증가)",
            "example": 1
          },
          "userId": {
            "type": "integer",
            "description": "소유 사용자 ID",
            "example": 1
          },
          "name": {
            "type": "string",
            "maxLength": 20,
            "description": "카테고리 이름 (동일 사용자 내 중복 불가, 최대 20자)",
            "example": "업무"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "카테고리 생성 일시 (ISO 8601, UTC)",
            "example": "2026-04-28T00:00:00.000Z"
          }
        },
        "required": ["id", "userId", "name", "createdAt"]
      },
      "Todo": {
        "type": "object",
        "description": "할일 정보. `status` 필드는 서버에서 동적으로 계산되어 응답에 포함됩니다.",
        "properties": {
          "id": {
            "type": "integer",
            "description": "할일 고유 식별자 (자동 증가)",
            "example": 1
          },
          "userId": {
            "type": "integer",
            "description": "소유 사용자 ID",
            "example": 1
          },
          "categoryId": {
            "type": "integer",
            "nullable": true,
            "description": "소속 카테고리 ID. null이면 미분류",
            "example": 1
          },
          "title": {
            "type": "string",
            "maxLength": 100,
            "description": "할일 제목 (필수, 최대 100자)",
            "example": "API 설계 완료하기"
          },
          "description": {
            "type": "string",
            "maxLength": 500,
            "nullable": true,
            "description": "할일 설명 (선택, 최대 500자)",
            "example": "OpenAPI 3.0 스펙 작성"
          },
          "dueDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "종료 목표 일시 (선택, ISO 8601, UTC). null이면 기한 초과 판별 대상에서 제외",
            "example": "2026-04-30T18:00:00.000Z"
          },
          "isCompleted": {
            "type": "boolean",
            "description": "완료 여부 (기본값: false)",
            "example": false
          },
          "status": {
            "$ref": "#/components/schemas/TodoStatus"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "할일 생성 일시 (ISO 8601, UTC)",
            "example": "2026-04-28T00:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "할일 최종 수정 일시 (ISO 8601, UTC). 수정 시 자동 갱신",
            "example": "2026-04-28T00:00:00.000Z"
          }
        },
        "required": ["id", "userId", "categoryId", "title", "description", "dueDate", "isCompleted", "status", "createdAt", "updatedAt"]
      },
      "TodoStatus": {
        "type": "string",
        "enum": ["pending", "completed", "overdue"],
        "description": "할일 상태 (동적 계산).\n- `pending`: 미완료 (`is_completed=false`, `due_date IS NULL` 또는 `due_date >= NOW()`)\n- `completed`: 완료 (`is_completed=true`)\n- `overdue`: 기한 초과 (`is_completed=false`, `due_date < NOW()`)",
        "example": "pending"
      },
...
      "CategoryCreateRequest": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 20,
            "description": "카테고리 이름 (필수, 최대 20자, 동일 사용자 내 중복 불가)",
            "example": "업무"
          }
        }
      },
      "TodoCreateRequest": {
        "type": "object",
        "required": ["title"],
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "description": "할일 제목 (필수, 최대 100자)",
            "example": "API 설계 완료하기"
          },
          "description": {
            "type": "string",
            "maxLength": 500,
            "nullable": true,
            "description": "할일 설명 (선택, 최대 500자)",
            "example": "OpenAPI 3.0 스펙 작성"
          },
          "categoryId": {
            "type": "integer",
            "nullable": true,
            "description": "소속 카테고리 ID (선택, 본인 소유 카테고리만 허용). null이면 미분류",
            "example": 1
          },
          "dueDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "종료 목표 일시 (선택, ISO 8601, UTC)",
            "example": "2026-04-30T18:00:00.000Z"
          }
        }
      },
      "TodoUpdateRequest": {
        "type": "object",
        "description": "할일 수정 요청. 포함된 필드만 업데이트됩니다. 최소 하나의 필드가 필요합니다.",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "description": "할일 제목 (최대 100자)",
            "example": "API 설계 및 문서화 완료하기"
          },
          "description": {
            "type": "string",
            "maxLength": 500,
            "nullable": true,
            "description": "할일 설명 (최대 500자). null 전달 시 설명 제거",
            "example": "OpenAPI 3.0 스펙 작성 및 Swagger UI 연동"
          },
          "categoryId": {
            "type": "integer",
            "nullable": true,
            "description": "소속 카테고리 ID. null 전달 시 미분류로 변경",
            "example": 1
          },
          "dueDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "종료 목표 일시 (ISO 8601, UTC). null 전달 시 기한 제거",
            "example": "2026-04-30T20:00:00.000Z"
          },
          "isCompleted": {
            "type": "boolean",
            "description": "완료 여부",
            "example": true
          }
        }
      },
      "SuccessResponse": {
        "type": "object",
        "description": "API 성공 응답 공통 포맷",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [true],
            "description": "요청 성공 여부 (항상 true)",
            "example": true
          },
          "data": {
            "description": "응답 데이터. 엔드포인트마다 구조가 다릅니다.",
            "nullable": true
          }
        },
        "required": ["success", "data"]
      },
      "ErrorResponse": {
        "type": "object",
        "description": "API 오류 응답 공통 포맷",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [false],
            "description": "요청 성공 여부 (항상 false)",
            "example": false
          },
          "message": {
            "type": "string",
            "description": "사람이 읽을 수 있는 오류 메시지",
            "example": "요청을 처리하는 중 오류가 발생했습니다."
          }
        },
        "required": ["success", "message"]
      }
    },
    "parameters": {
      "TodoId": {
        "name": "id",
        "in": "path",
        "description": "할일 고유 식별자",
        "required": true,
        "schema": {
          "type": "integer",
          "minimum": 1
        },
        "example": 1
      },
      "CategoryId": {
        "name": "id",
        "in": "path",
        "description": "카테고리 고유 식별자",
        "required": true,
        "schema": {
          "type": "integer",
          "minimum": 1
        },
        "example": 1
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "인증 실패. JWT 토큰이 없거나 유효하지 않거나 만료되었습니다.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "success": false,
              "message": "인증이 필요합니다. 로그인 후 다시 시도해주세요."
            }
          }
        }
      },
...
      "TodoNotFound": {
        "description": "할일을 찾을 수 없음",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "success": false,
              "message": "존재하지 않는 할일입니다."
            }
          }
        }
      },
      "CategoryNotFound": {
        "description": "카테고리를 찾을 수 없음",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "success": false,
              "message": "존재하지 않는 카테고리입니다."
            }
          }
        }
      },
      "InternalServerError": {
        "description": "서버 내부 오류",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "success": false,
              "message": "서버 내부 오류가 발생했습니다. 잠시 후 다시 시도해주세요."
            }
          }
        }
      }
    }
  }
}